+++ Model/Api/Builder/OrderRequestBuilder/GatewayInfoBuilder/In3GatewayInfoBuilder.php
@@ -1,0 +1,69 @@
+<?php
+/**
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Open Software License (OSL 3.0)
+ * that is provided with Magento in the file LICENSE.txt.
+ * It is also available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/osl-3.0.php
+ *
+ * Copyright © 2020 MultiSafepay, Inc. All rights reserved.
+ * See DISCLAIMER.md for disclaimer details.
+ *
+ */
+
+declare(strict_types=1);
+
+namespace MultiSafepay\ConnectCore\Model\Api\Builder\OrderRequestBuilder\GatewayInfoBuilder;
+
+use Magento\Framework\Exception\LocalizedException;
+use Magento\Sales\Api\Data\OrderInterface;
+use Magento\Sales\Api\Data\OrderPaymentInterface;
+use MultiSafepay\Api\Transactions\OrderRequest\Arguments\GatewayInfo\Meta;
+use MultiSafepay\ValueObject\Customer\EmailAddress;
+use MultiSafepay\ValueObject\Customer\PhoneNumber;
+use MultiSafepay\ValueObject\Date;
+use MultiSafepay\ValueObject\Gender;
+
+class In3GatewayInfoBuilder
+{
+    /**
+     * @var Meta
+     */
+    private $meta;
+
+    /**
+     * GatewayInfo constructor.
+     *
+     * @param Meta $meta
+     */
+    public function __construct(
+        Meta $meta
+    ) {
+        $this->meta = $meta;
+    }
+
+    /**
+     * @param OrderInterface $order
+     * @param OrderPaymentInterface $payment
+     * @return Meta
+     * @throws LocalizedException
+     */
+    public function build(OrderInterface $order, OrderPaymentInterface $payment): Meta
+    {
+        $billingAddress = $order->getBillingAddress();
+
+        if ($billingAddress === null) {
+            $msg = __('The transaction could not be created because the billing address is missing');
+            throw new LocalizedException($msg);
+        }
+
+        $additionalInformation = $payment->getAdditionalInformation();
+
+        return $this->meta->addGender(new Gender($additionalInformation['gender']))
+            ->addBirthday(new Date($additionalInformation['date_of_birth']))
+            ->addEmailAddress(new EmailAddress($order->getCustomerEmail()))
+            ->addPhone(new PhoneNumber($billingAddress->getTelephone()));
+    }
+}